retrain_model.py
from __future__ import division
from __future__ import print_function
from __future__ import absolute_import

import os
import io
import pandas as pd
import tensorflow as tf

from PIL import Image
from object_detection.utils import dataset_util
from collections import namedtuple, OrderedDict
import argparse

new_model = tf.keras.models.load_model('./model')

# Check its architecture
new_model.summary()