requirements.txt
tensorflow==1.14
tensorflow-hub
pillow==6.2.1
lxml==4.4.1
jupyter==1.0.0
matplotlib==3.1.1
opencv-python
numpy
object_detection
